App/requirements.txt
discord.py-self
discord-webhook